remove Curl from SDK codes.
