+++ KiiThingSDK/kii_prv_utils.c
@@ -98,25 +98,6 @@ static char* prv_url_encoded_copy(char* s1, const char* s2)
     return kii_strncpy(s1, s2, kii_strlen(s2)+1);
 }
 
-struct curl_slist* prv_curl_slist_create(const char* first, ...)
-{
-    struct curl_slist* retval = NULL;
-    const char* header = NULL;
-    va_list list;
-    va_start(list, first);
-    for (header = first; header != NULL; header = va_arg(list, char*)) {
-        struct curl_slist* tmp = curl_slist_append(retval, header);
-        if (tmp == NULL) {
-            curl_slist_free_all(retval);
-            retval = NULL;
-            break;
-        }
-        retval = tmp;
-    }
-    va_end(list);
-    return retval;
-}
-
 kii_char_t* prv_new_header_string(const kii_char_t* key,
                                   const kii_char_t* value)
 {
@@ -161,87 +142,6 @@ kii_char_t* prv_new_auth_header_string(const kii_char_t* access_token)
     kii_strncpy(ret, authbearer, authbearerLen + 1);
     kii_strncat(ret, access_token, tokenLen + 1);
     return ret;
-}
-
-struct curl_slist*
-prv_common_request_headers(
-                           const kii_app_t app,
-                           const kii_char_t* opt_access_token,
-                           const kii_char_t* opt_content_type)
-{
-    kii_char_t* app_id_hdr = NULL;
-    kii_char_t* app_key_hdr = NULL;
-    struct curl_slist* retval = NULL;
-    
-    M_KII_ASSERT(app->app_id != NULL);
-    M_KII_ASSERT(app->app_key != NULL);
-    
-    app_id_hdr = prv_new_header_string("x-kii-appid", app->app_id);
-    app_key_hdr = prv_new_header_string("x-kii-appkey", app->app_key);
-    if (app_id_hdr == NULL || app_key_hdr == NULL) {
-        retval = NULL;
-        goto ON_EXIT;
-    }
-    
-    retval = prv_curl_slist_create(app_id_hdr, app_key_hdr, NULL);
-    if (retval == NULL) {
-        goto ON_EXIT;
-    }
-    
-    if (opt_access_token != NULL) {
-        kii_char_t* auth_hdr = prv_new_auth_header_string(opt_access_token);
-        struct curl_slist* tmp = (auth_hdr == NULL) ? NULL :
-        curl_slist_append(retval, auth_hdr);
-        
-        kii_dispose_kii_char(auth_hdr);
-        if (tmp == NULL) {
-            curl_slist_free_all(retval);
-            retval = NULL;
-            goto ON_EXIT;
-        }
-        retval = tmp;
-    }
-    
-    if (opt_content_type != NULL) {
-        kii_char_t* content_type_hdr = prv_new_header_string("content-type",
-                                                             opt_content_type);
-        struct curl_slist* tmp = (content_type_hdr == NULL) ? NULL :
-        curl_slist_append(retval, content_type_hdr);
-        
-        kii_dispose_kii_char(content_type_hdr);
-        if (tmp == NULL) {
-            curl_slist_free_all(retval);
-            retval = NULL;
-            goto ON_EXIT;
-        }
-        retval = tmp;
-    }
-    
-ON_EXIT:
-    M_KII_FREE_NULLIFY(app_id_hdr);
-    M_KII_FREE_NULLIFY(app_key_hdr);
-    
-    return retval;
-}
-
-struct curl_slist*
-prv_curl_slist_append_key_and_value(struct curl_slist *headers,
-                                    const kii_char_t *key,
-                                    const kii_char_t *value)
-{
-    kii_char_t* hdr = prv_new_header_string(key, value);
-    struct curl_slist* retval = (hdr == NULL) ? NULL :
-    curl_slist_append(headers, hdr);
-    kii_dispose_kii_char(hdr);
-    return retval;
-}
-
-void prv_log_req_heder(struct curl_slist* header)
-{
-    while (header != NULL) {
-        prv_log("req header: %s", header->data);
-        header = header->next;
-    }
 }
 
 int prv_log(const char* format, ...)
