+++ KiiThingSDK/kii_prv_utils.h
@@ -10,12 +10,6 @@
 
 #include <stdio.h>
 
-#ifdef XCODE
-#include "curl.h"
-#else
-#include <curl/curl.h>
-#endif
-
 #include "kii_custom.h"
 
 #ifdef __cplusplus
@@ -37,32 +31,10 @@ void prv_kii_set_info_in_error(
 /* Returned value of this method must be freed by caller of this method. */
 char* prv_build_url(const char* first, ...);
 
-/* The last argument of this method must be NULL.
-
-   If all headers are appended, then this method returns struct
-   curl_slist ponter. Otherwise returns NULL.
-
-   Returned value must be freed by caller of this method with
-   curl_slist_free_all. */
-struct curl_slist* prv_curl_slist_create(const char* first, ...);
-
 kii_char_t* prv_new_header_string(const kii_char_t* key,
                                   const kii_char_t* value);
 
 kii_char_t* prv_new_auth_header_string(const kii_char_t* access_token);
-
-struct curl_slist*
-prv_common_request_headers(
-                           const kii_app_t app,
-                           const kii_char_t* opt_access_token,
-                           const kii_char_t* opt_content_type);
-
-struct curl_slist*
-prv_curl_slist_append_key_and_value(struct curl_slist* headers,
-                                    const kii_char_t* key,
-                                    const kii_char_t* value);
-
-void prv_log_req_heder(struct curl_slist* header);
 
 int prv_log(const char* format, ...);
 int prv_log_no_LF(const char* format, ...);
